requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
soundfile==0.13.1
pydantic==2.11.7
requests==2.32.5
huggingface_hub==0.34.4

vastai-sdk
git+https://github.com/OpenMOSS/MOSS-TTS.git